Move db-migrate packages (#291)
https://github.com/DEFRA/water-abstraction-team/issues/36

Whilst we were trying to bring some sanity to the deployment process we found that the db-migrate package had been installed as a dev dependency. This dependency is required as part of the deployment process which has meant currently ALL dependencies have to be installed upon deployment, the devDependencies cannot be ignored.