Remove 'local' from NODE_ENV example
To be honest, we making this minor change to test a new deployment script. However, it's still [not advised](https://stackoverflow.com/a/16979503/6117745) to set `NODE_ENV` to anything other than `development` or `production` and the new dev team are already falling foul of issues caused by setting it to other values. 

Anyway, to avoid using it in this way we are updating the .env.example though really, we just need a minor tweak for testing purposes!