Remove last use of NODE_ENV
Looks like we were giving our local dev environments a boost when it came to connecting to postgresql. We don't agree with making adjustments like this for local development. Plus, based on what we've seen so far we're not confident we should be playing with the efault postgres connection config at all.

For now, we'll settle on just using the same config no matter the environment and leave changes to PG config to a wider review of the projects.