Refactor HapiPinoIgnoreRequestService as module (#324)
We copied this from [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) just prior to use refactoring our use of classes.

We managed to catch the other repos where we've been replacing **hapi/good* before the change got merged. But this one got through.

So, just a quick change to refactor the service to be a module rather than a class.