Only log errors (#337)
https://github.com/DEFRA/water-abstraction-team/issues/79

We have an ongoing problem where the log files for our apps very quickly become massive. In the space of a week, we can see a repo like [water-abstraction-returns](https://github.com/DEFRA/water-abstraction-returns) generate a log file almost 10Gb in size!

Fortunately, we have a solution in place for `production` that truncates them on a regular basis. But because we switch off our non-prod environments that process is often blocked there, meaning the logs get out of hand and use up all available disk space. It also means our local environments can be extremely noisy and if we don't switch them off, they too can run into the same issue.

So, for our background web services, we are moving to only logging errors in the hope this reduces the size of our logs. The change only applies to what [hapi-pino](https://github.com/pinojs/hapi-pino) is doing. So, any logging we do within the app code will still be emitted. But when a service like **water-abstraction-returns**, gets hammered with requests we'll be removing a lot of unnecessary output.