Removing console.log() from legacy migrations (#374)
* Removing console.log() from legacy migrations

https://github.com/DEFRA/water-abstraction-team/issues/91

This PR is focusing on removing console.log() from legacy migrations as per Teams issue. When working on DEFRA/water-abstraction-service#2227 we again broke GitHub CI because of the volume of output from our migrations.

In DEFRA/water-abstraction-team#65 we thought we'd dealt with the issue believing that db-mgrate's verbose() argument was to blame.

Looking into the issue in the stuck bill run we came across a db-migrate/node-db-migrate#453 (comment) that gave us a 🤦 !

The db-migrate template used when generating migrations includes a console.log() that outputs the migration file read in. So, though we might not be outputting all the SQL being fired, we are still outputting the contents of every single migration file!

We definitely don't need to see this and it should remove the chance of us breaking the build in this way once and for all.

So, this issue is about going into the existing migrations and removing the 2 console.log('received data: ' + data) lines in each across all repos.

* Remove console.log() from legacy migrations