Add `notes` and `created_by` to `versions` table (#406)
https://eaflood.atlassian.net/browse/WATER-4807

New functionality is being created to enable the user to save a note against a return submission, previously referred to as a version.

This PR will add the `notes` and `created_by` columns to the `returns.versions` table. A subsequent PR will make similar changes to our views and test data migrations in the system repo.