Create historic return cycles (#417)
https://eaflood.atlassian.net/browse/WATER-5007

As part of our work to replace NALD as the main source of managing returns we need to be able to handle not only new returns but historic changes to returns. This PR adds a migration to create historic return cycles so if any changes are made on very old licences we will correctly generate the necessary returns logs.