Add sent date column
https://eaflood.atlassian.net/browse/WATER-5018

WRLS is taking over management of the 'returns leg' from NALD. Contingent to that is the ability to continue reporting on returns.

Our work with RDP is behind the changes to the service, but we need to make the switch soon in order to support the first quarterly return submissions in July. So, we're going to ensure we can generic the same reports by creating some ad-hoc queries.

Straight off we've spotted that the existing BOXI reports include `send_date`. This is something the [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) doesn't currently import from NALD but could with a minor change. However, the first step is to add a migration to add a field to the `returns.returns` table.