Fix missing received date in return log (#425)
https://eaflood.atlassian.net/browse/WATER-5075

Whilst working on creating a new 'clean' empty VOID return logs job, we found an edge case in the data.

One of the return logs has submitted returns data, but its `received_date` is missing.

We believe this is a residual record that was affected by the issue in WATER-5057. There was a point that the timings of the 'FME WRLS to NALD' and the 'FME/WRLS NALD to WRLS' imports were out of sync, which led to return logs getting their `received_date` blanked.

To fix it, the dates needed to be manually re-entered into NALD, and this one appears to have been missed.