Fix corrupt return logs for a licence
https://eaflood.atlassian.net/browse/WATER-5137

The business reported that when viewing the return logs for a licence, the UI was not making sense.

One was showing as received, when it is not actually due yet. Another was showing as 'due' in November, but when you click into it there is submission data.

They were correct that the data was wrong for these return logs. The source of the confusion was multiple NALD return logs for the same period. This is not a scenario the import was built to handle, and from discussions is an exception in NALD.

These means the import has pulled data from multiple sources when trying to setup the single return log in WRLS.

The returns import is now disabled as WRLS has taken over management of all returns from NALD. So, the only solution is a data fix migration for this licence.