Fix corrupt return logs for a licence (#430)
https://eaflood.atlassian.net/browse/WATER-5137

The business reported that when viewing the return logs for a licence, the UI was not making sense.

One was showing as received when it is not actually due yet. Another was showing as 'due' in November, but when you click into it, there is submission data.

They were correct that the data was incorrect for these return logs. The source of the confusion was multiple NALD return logs for the same period. This is not a scenario the import was designed to handle, and from discussions, it appears to be an exception in NALD.

This means the import has pulled data from multiple sources when trying to set up the single return log in WRLS.

The returns import is now disabled as WRLS has taken over management of all returns from NALD. So, the only solution is a data fix migration for this licence.