Allow null due date on returns table (#436)
https://eaflood.atlassian.net/browse/WATER-5212

As part of our work to get ready for new regulations we need to leave the due date out of the return log when created. This due date will be sent at a later time once we have confirmed we have sent the notification.

This PR updates the database column to set it to null.