Return cycle table allow null due date (#437)
https://eaflood.atlassian.net/browse/WATER-5212

As part of our work to get ready for new regulations we need to leave the due date out of the return logs when created. As this due date will be set after the users have been notified we no longer need to have a due date on a return cycle.

This PR updates the database column to set it to null.