Fix legacy migration
Our new migration depends on the relatively new column `quarterly` in `returns.returns`. However, it failed when ran in CI because the column doesn't exist.

Huh?!

We tracked it down to the fact that the migration which added the column was also populating it with data from another schema. When this happens we have to wrap the part of a query which references other schemas in a `DO` block, essentially an anonymous function. Else, it will fail in CI.

What we didn't spot when this was added is that we'd also included the adding of the new column inside the `DO` block. This was safe to execute outside of it, and actually needs to execute outside of it for later migrations that depend on the new `quarterly` column to work.