Round return quantities to 6 decimals (#455)
https://eaflood.atlassian.net/browse/WATER-5394

When return volumes are entered and calculated we do some maths that can lead to very long decimal points when saved into the database. This can lead to inconsistent values being shown to users depending on the numbers entered and conversions between different volume units.

This change adds a rounding when the quantity is saved to the database.